requirements.txt
numpy==1.19.2
opencv-python==4.4.0.42
object_detection_api/object_detection-0.1-py3-none-any.whl